loading page fix????

+++ hwbounty-home/src/LoadIntoCache.js
@@ -2,46 +2,52 @@ import axios from "axios";
 import { Component } from "react";
 import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
-
-class LoadIntoCache extends Component{
-	constructor(props){
+const sleep = (delay) => new Promise((resolve) => setTimeout(resolve, delay));
+class LoadIntoCache extends Component {
+	constructor(props) {
 		super(props);
 	}
-	async updateUser(){
+	async updateUser(tries, delay) {
+		await sleep(delay || 0);
 		let data = await axios.get("https://api.hwbounty.help/@me").catch(console.trace);
-		if (data && data.status === 200 && data.data){
-			localStorage.setItem("user",JSON.stringify(data.data));
-		}else{
-			localStorage.clear();
+		if (data && data.status === 200 && data.data) {
+			localStorage.setItem("user", JSON.stringify(data.data));
+		} else {
+			if (!tries || tries < 3)
+				await this.updateUser(tries ? tries + 1 : 1, 5000 * tries);
+			else {
+				localStorage.clear();
+			}
+
 		}
 	}
-	async updateSchedule(){
-		axios.get("https://api.hwbounty.help/schedule/@me").then(data=>{
-			if (data && data.status === 200){
-				localStorage.setItem("cachedSchedule",JSON.stringify(data.data));
+	async updateSchedule() {
+		axios.get("https://api.hwbounty.help/schedule/@me").then(data => {
+			if (data && data.status === 200) {
+				localStorage.setItem("cachedSchedule", JSON.stringify(data.data));
 			}
 		}).catch(console.trace);
-		axios.get("https://api.hwbounty.help/sgy/getZoomLinks").then(data=>{
-			if (data && data.status === 200){
-				localStorage.setItem("cachedCourseInfo",JSON.stringify(data.data));
+		axios.get("https://api.hwbounty.help/sgy/getZoomLinks").then(data => {
+			if (data && data.status === 200) {
+				localStorage.setItem("cachedCourseInfo", JSON.stringify(data.data));
 			}
 		})
 
 	}
-	componentDidMount(){
+	componentDidMount() {
 		let oldPageUrl = "";
-		setInterval(()=>{
-			if (oldPageUrl !== this.props.location.pathname){
+		setInterval(() => {
+			if (oldPageUrl !== this.props.location.pathname) {
 				oldPageUrl = this.props.location.pathname;
 				this.updateCache();
 			}
-		},1000);
+		}, 1000);
 	}
-	async updateCache(){
+	async updateCache() {
 		this.updateUser();
 		this.updateSchedule();
 	}
-	render(){
+	render() {
 		// this.updateCache();
 		return null;
 	}
